+++ src/layout/dagre/src/order/barycenter.ts
@@ -9,7 +9,7 @@ const barycenter = (g: Graph, movable: string[]) => {
       return { v };
     }  {
       const result = { sum: 0, weight: 0 };
-      inV.forEach((e) => {
+      inV?.forEach((e) => {
         const edge = g.edge(e);
         const nodeU = g.node(e.v);
         result.sum += (edge.weight * (nodeU.order as number));