+++ src/layout/dagre/src/order/index.ts
@@ -80,7 +80,7 @@ const sweepLayerGraphs = (layerGraphs: IGraph[], biasRight: any, usePrev?: any) 
   layerGraphs?.forEach((lg) => {
     const root = lg.graph().root as string;
     const sorted = sortSubgraph(lg, root, cg, biasRight, usePrev);
-    sorted.vs.forEach((v: string, i: number) => {
+    sorted.vs?.forEach((v: string, i: number) => {
       lg.node(v).order = i;
     });
     addSubgraphConstraints(lg, cg, sorted.vs);
@@ -89,7 +89,7 @@ const sweepLayerGraphs = (layerGraphs: IGraph[], biasRight: any, usePrev?: any) 
 
 const assignOrder = (g: IGraph, layering: any) => {
   layering?.forEach((layer: any) => {
-    layer.forEach((v: string, i: number) => {
+    layer?.forEach((v: string, i: number) => {
       g.node(v).order = i;
     });
   });